更新配置和类型定义以支持 Windows 环境
- 修改 `.editorconfig` 文件以使用 CRLF 换行符。
- 在 `font.h` 中简化 `FONT_FAMILY` 类型声明。
- 在 `time_window.h` 中将 `StepCountDown` 的类型从 `UINT` 改为 `long long`，以支持更大数值。
- 移除 `tray.h` 中的 `namespace Tray` 定义以简化代码结构。

+++ time_window.h
@@ -49,7 +49,7 @@ int CreateTimeClassAndWindow(_In_ HINSTANCE hInstance, _In_ int nCmdShow);
 class StepCountDown {
 
 public:
-    StepCountDown(UINT down, UINT count = 0) : count(count), down(down) {}
+    StepCountDown(long long down, long long count = 0) : count(count), down(down) {}
     StepCountDown(std::chrono::seconds down, std::chrono::seconds count = std::chrono::seconds(0)) : count(count.count()), down(down.count()) {}
 
     StepCountDown() : StepCountDown(0, 0) {};
@@ -58,22 +58,22 @@ public:
         this->count += i;
     }
 
-    UINT distance() const {
+    long long distance() const {
         return abs(static_cast< int >( this->down - this->count ));
     }
 
-    void reset(UINT end, UINT start = 0) {
+    void reset(long long end, long long start = 0) {
         this->count = start;
         this->down = end;
     }
 
-    void reset(std::chrono::seconds end, UINT start = 0) {
+    void reset(std::chrono::seconds end, long long start = 0) {
         this->count = start;
         this->down = end.count();
     }
 
 private:
 
-    UINT count;
-    UINT down;
+    long long count;
+    long long down;
 };
